+++ spec/spec_helper.rb
@@ -9,6 +9,7 @@ require 'fue'
 require 'English'
 require 'webmock/rspec'
 require 'graphlient'
+require 'recursive-open-struct'
 
 RSpec.configure(&:raise_errors_for_deprecations!)