+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.3 (Next)
 
+* [#6](https://github.com/dblock/fue/issues/6): Configure max depth and breadth separately - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.2 (Next)